+++ src/main/kotlin/edu/jimei/janus/controller/NotificationController.kt
@@ -6,6 +6,7 @@ import edu.jimei.janus.domain.notification.NotificationType
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
+import java.math.BigDecimal
 import java.util.UUID
 
 @RestController
@@ -166,7 +167,7 @@ class NotificationController(
     @PostMapping("/grade/{submissionId}/published")
     fun notifyGradePublished(
         @PathVariable submissionId: UUID,
-        @RequestParam score: Double,
+        @RequestParam score: BigDecimal,
         @RequestParam studentId: UUID,
         @RequestParam teacherId: UUID
     ): ResponseEntity<NotificationDto> {
